+++ lib/components/SelectionControl/checkbox/checkbox-single/checkbox.scss
